reset date button, simplified addEvent(), refactored EventListView, reindenting
